[FIX] delivery_easypost: round up the quantity
When trying to deliver a product with EasyPost, if the quantity is
decimal, it will lead to an error message:
> Easypost returned an error: Wrong parameter type.
> - quantity: must be a whole number

According to EasyPost Documentation, `CustomsItem.quantity` should be a
float. However, considering the EasyPost server response and the
confirmation from EasyPost Support:
> The value for this field as the error message says, needs to be a
whole number, there cannot be a decimal in it. I can make a note of the
option type saying "float" to pass on to correct the documentation.

`quantity` should actually be a number

OPW-2619670

closes odoo/enterprise#22104

X-original-commit: 6b25ff6ffcfc712fda6de4c121086315e4f7db91
Signed-off-by: Arnold Moyaux <arm@odoo.com>
Signed-off-by: Adrien Widart <awt@odoo.com>