[FIX] helpdesk: SLAs are unarchived unexpectedly
When helpdesk.team record is updated, _check_sla_group was triggered
everytime, which caused to force enable/disable all SLAs.

It is incorrect to trigger it every time on helpdesk team update,
because there could be some archived SLAs that are not intended to be
used, yet it would be unarchived and would mess up new tickets, where
old SLA could take priority.

It should be only triggered if we know what change was related with
enabling/disabling SLA feature.

opw:2658436

closes odoo/enterprise#22114

X-original-commit: e064f3f496f15f8ddaa5e15a1440d443b96de805
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>
Co-authored-by: oerp-odoo <andrius@focusate.eu>