[FIX] website_calendar_crm: current user don't need to be follower
Issue
-----
By default the create_uid is insert as followers.
When a lead is created from an online appointement we don't need this
user as followers since most of the time the current user is the public
user.

This is harmless until someone is wondering why the public user do not
received the email sent to follower and change the email of the public
user.

Solution
--------
Do not add self.env.user as follower on the lead created.
If a real user create the appointement and thus the lead
he will be suggested the first time someone send an email.

closes odoo/enterprise#22063

X-original-commit: b97b361789f1a86d7db06058f2d471b46aa329fd
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>