[IMP] crm_helpdesk: propagate UTMs when converting tickets & leads
Information about UTM codes was lost when converting leads to
helpdesk tickets and vice versa. This commit ensures that they
are correctly propagated.

The Python test suite is completed to reflect this.

As specific processes are in place to create Helpdesk Teams
and Ticket Types, they can no longer be created nor edited
on the fly in this conversion process.
To accelerate form completion in case there is a single
helpdesk team to chose from, this field will be
automatically completed on the Lead to Ticket wizard form.

Task-2671664

closes odoo/enterprise#21848

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>