[FIX] stock_barcode: scan tracked product
Before this commit, if the user scans a tracked product barcode and
there is already an existing quant for this product, it won't find
existing quant if it has no lot/serial number.

How to reproduce:
  - Create a product tracked by serial number and set a barcode;
  - Go to the Inventory Adjustments and create a quant for this product,
    set an `inventory_quantity` but no serial number and apply;
  - Go to Barcode > Inventory Adjustments and scan the product barcode
    --> A line for this product is created but the qty. set by the
        previous Inv. Adj. isn't set on the line.

Also fix a little issue with the Set Button which happens in this case:
as the max qty for a product tracked by SN is supposed to be 1, the
button didn't work properly if the line `quantity` is greater than 1,
so the line inv. qty can only be switched to 0 and 1 and the line could
not be marked as unset.

task-2682383

X-original-commit: ba9d439f0222408d1a15b857ec65d2a8945b38b9
Part-of: odoo/enterprise#22989