[IMP] helpdesk(_*): improve the user/validation errors for helpdesk
Currently, the user/validation errors are sometimes not correct in
English and/or quite obscure and don't help the user understand/solve
the issue.

so some copywriting should help make the experience better.
In addition, some errors could also be avoided beforehand to provide
a better ux.

In this commit so in this commit removed several constraint and improved
user/validation messages.

task-2615468

closes odoo/enterprise#22192

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>