[FW][IMP] delivery_easypost: better user error #21865
Before this commit, when the client meets an issue with EasyPost, they get an
uncomplete explanation.

For example when EasyPost gives this response:
`{"error":{"code":"PARAMETER.INVALID_TYPE","message":"Wrong parameter type.","errors":[{"field":"quantity","message":"must be greater than or equal to 1"}]}}`
Odoo raised the UserError containing only the `message`: "Wrong parameter type".
With this information, the client does not have any idea of what is wrong.
Following the easypost API, We can see that the error object is not just made
of the message (https://www.easypost.com/docs/api#error-object).
With the content of the `errors` field, this is way clearer.

opw-2619670

closes odoo/enterprise#22113

Forward-port-of: #21758
X-original-commit: 0412a8b4955809f04bc006f70cf75e45f813219b
Signed-off-by: Arnold Moyaux <arm@odoo.com>