[FIX] social_facebook: Show error when no page linked to account
Before this commit, when trying to add a linkedin account
with no page linked, there was no error, nothing was shown.

Now, an explicit error is shown to the user.

Steps to reproduce:
- Have a V14 with social
- Have a Facebook account
- Go to Marketing Social
    - Add a Stream
        - Choose Facebook
	- Don't check any pages when Facebook ask for it and continue
        - Connect your account

After theses steps, the browser comes back to the Social
Marketing page but nothing has changed and the `social.account`
is not created.

closes odoo/enterprise#22629

X-original-commit: fb94b61c4956d7d6d8ff4d5635fbef7ea371bd8a
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>