[FIX] l10n_ar_edi: AFIP message format error + do not show hints
If there is an observation or event we are posting a message in the
invoice chatter with the detail information,

The problem is that the message is in HTML and we are using \n as
break lines not <br/>: using plaintext2html odoo function solve the
problem to show the message in the proper format.

No hints:
Sometimes when a big change in AFIP is comming they send an event info
in the response we receive from validating an invoice. This events has a
message and a code.

We were interpreting the event code in tht HINT sections as if was an
obervation code and this is wrong and confusing for the user.

Now we are simply not taking into account the events code in the HINT
section, this way the HINT section will only reflect help if there is
any error or observation for AFIP when validating the invoice.

closes odoo/enterprise#22278

X-original-commit: 0f08cb70ec8bb7d29823ef78c1421982a0571643
Signed-off-by: Josse Colpaert <jco@openerp.com>