[FIX] sale_ebay: fix the UOM change
Steps to reproduce the bug:
- Install sale_ebay
- Go to inventory settings > Enable “Units of Measure” option
- Create a new storable product > save
- Edit > Change the Uom
- Traceback is triggered

Problem:
When ebay is installed, there is a list on the form of a product template. This list displays all variants and some of their fields.
One of these is “virtual_available” which is a computed field. When changing the UoM of the product, the _compute_quantitie method will be triggered.
However, since the UoM isn't defined in the tree view of the variants, this field will be false during the computations.
This will lead to an error as the UoM is required for the compute method to be correctly processed: https://github.com/odoo/odoo/blob/15.0/addons/stock/models/product.py#L186

opw-2669079

closes odoo/enterprise#22099

X-original-commit: f404857a3aa3a649c74cee53afd2bcaf55d70380
Signed-off-by: William Henrotin (whe) <whe@odoo.com>
Signed-off-by: Djamel Touati (otd) <otd@odoo.com>