[IMP] sign: remove sms button from role form view, add it in settings
In the role form view there is, currently, a button for adding sms credits to
the account. However, the role form view is never accessed, since the tree view
is directly editable. This PR removes the button from the form view and adds it
to the settings page of sign.

task-2679754

closes odoo/enterprise#22120

Signed-off-by: Arnaud Joset <arj@odoo.com>