[FIX] account: collapse journal list in financial reports
If numerous journals were selected, their names could end up covering options

Now : when a journal group is selected, its name is displayed in the search field.

When multiple journals are selected or when none are selected, All Journals is displayed in the search field.

From five journals selected, the displayed name is completed by 'and X other(s)'
The filter allowing for the journal groups to be applied is now located in the init_filter_journals function.

task-2671733

closes odoo/enterprise#23140

X-original-commit: c508fae8af1ab3a2417e7d2b4bc917a4c6f4a47d
Signed-off-by: Laurent Smet <las@odoo.com>