[FIX] mass_mailing_themes: remove superfluous and wrong xml attributes
Mass mailing themes had a number of superfluous and wrong xml attributes
typically copied from the DOM but not supposed to be found in templates.
It was generally without consequence except in the case of the VIP theme
where two images had the "o_modified_image_to_save" class, eventually
causing a traceback on save.

task-2680428

closes odoo/enterprise#22189

X-original-commit: d0644b725215355698bdb2bd15a7a54a691864f0
Signed-off-by: David Monjoie (dmo) <dmo@odoo.com>
Signed-off-by: Antoine Guenet <age@odoo.com>