[FIX] sign: show/hide "Share" button correctly
Show the "Share" button in the template editing page when there is
only one role when the template is opened.

Before this commit:
In the template editing page. The "Share" button is displayed when
there are more than one sign.requests using this template, which is
meaningless and wrong.

Task-id: 2686174
Part-of: odoo/enterprise#22721