[IMP] l10n_nl_reports: add balance sheet respecting NL format
Currently, the NL localization uses the generic balance sheet, which is not correct according to NL government.
On top of that, the generic balance sheet is linked to the generic P&L, even though a specific P&L for NL has already been implemented.

In order to be more robust, a control domain on top of a domain adds a security layer that can help in debugging when adding new accounts.
This can be done by using account tags with control domains.

Changes:
- Add a balance sheet that respects the NL format.
- Rewrite the profit and loss report to make use of account codes and add control domain with tags.

Task id=2674603

closes odoo/enterprise#23115

X-original-commit: 1a2aba2b2538d8a495af2776f04a31adfa1cc31d
Related: odoo/odoo#81943
Signed-off-by: Laurent Smet <las@odoo.com>