[IMP] account_invoice_extract: allow the use of intracom taxes
Previously, the intracom taxes set on an invoice line during the
extraction process was always removed.
We now first remove it, check if the total amount is still the same
and if it is, we can re-apply it. It's a bit weird, but it avoids
re-implementing any tax computation mechanism.

closes odoo/enterprise#22226

X-original-commit: 81a2549f1c042b95a41abe8dd4876e49780ef8cc
Signed-off-by: Florian Daloze (fda) <fda@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>