[FIX] appointment: fix traceback while searching appointments
With commit[1], we added an appointment profile for the user but
while searching the appointments we are getting traceback.

With this task, we modify the def _get_search_domain() method to
def _get_appointment_search_domain() because sale_timesheet and
hr_timesheet modules have the same method of the controller of the portal
module so, while calling that method it also gets the domain result from
both modules and getting the traceback. (ValueError: Invalid field...)

commit[1] - 9b862fe

Task-2681283

closes odoo/enterprise#22504

X-original-commit: 1270c0e999e64dc032c79c1aed2b6037bc31b7d6
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>