[FIX] hr_contract_salary: Remove field from view created on hr_contract
The field contract_type_id is added in hr_contract, and now is added on
the view of that module on https://github.com/odoo/odoo/pull/79723

Why not a custom module to add this field in the view?
The hr_contract is the base module, then, my custom module installs it,
but it needs to set the contract type and the field is not present.

The field cannot be added in the custom module, because some customers
that use my module also needs hr_contract_salary, which adds the field in
the view.

closes odoo/enterprise#22248

X-original-commit: 3ecf389de430423dd3957119cc1b0f4017b496a6
Related: odoo/odoo#79763
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>