[IMP] project_enterprise,worksheet: improved error message and ux
Currently, User/validation errors are sometimes not correct in English
and/or quite obscure and don't help the user understand/solve the issue.
also there is constraint error on duplcation of the worksheet template

So in this commit, done some copywriting to make the experience better.
and inherit the view of demo data in proper place and hide the duplicate
option from the action of worksheet view.

Also there is validation error when task is marked as done but
delivery still pending
so in this commit allow to decrease the qty of product if delivery
is still pending.

task-261546

closes odoo/enterprise#23013

Related: odoo/odoo#81711
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>