[FIX] sale_subscription: Synchronize followers on partner modification
Expected Behavior

1. When making a new subscription, the client should be added only when validating
the subscription (aka move it into an in progress stage)
2. When we change the partner on a subscription, the related followers should be
updated too (Remove the old one and add the new one, if in progress).

Observed behaviour

From V13, everytime we change the client, the new one is added as a new follower,
while the old one stay as a follower, leading to sending mail to people not related
to the final subscription

Reproducibility

This bug can be reproduced following these steps:
1. Create a new subscription with a client
2. Set a new client
3. Edit the subscription, change the client and save
4. Check the followers

Problem Root Cause

The problem comes from the fact we set the client as follower each time we got through the
"write" method of the subscription (called whenever something is changed), without
regarding what as been changed and the current subscrption stage. Moreover, nothing was
made to unset an old client as a follower, even when changind the client ID. Finally,
the client ID was still editable even in a "in_progress" subscription.

Related issue
- opw-266755

closes odoo/enterprise#22361

X-original-commit: 114c0018ecf5f4672190dab2348bb8a5f4ecd479
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>