[FIX] l10n_cl_edi_boletas: Handle invalid token codes
This fix is almost the same as the one that was done for l10n_cl_edi in
https://github.com/odoo/enterprise/pull/21503 and https://github.com/odoo/enterprise/pull/20868 .
The logic is duplicated between modules and affected by the same issues.

Unlike _analyze_sii_result, _analyze_sii_sales_book_result and
_analyze_sii_result_rest consider all codes that weren't in their dicts
of known codes to indicate a rejected status. That means the user needs
to manually retry sending the daily sales book or boletas invoice in
case of an unexpected result. For the SOAP calls this has been updated
so that like regular electronic invoices and stock pickings, it will
raise an exception and log the unexpected response instead. The SII
status will remain on `ask_for_status` in that case, resulting in a
retry later. For the boletas REST call, the case will be dealt with in
https://github.com/odoo/enterprise/pull/22207 , since that code turned
out to have additional issues.

Because of the original logic, all potential rejected messages weren't
necessarily in the dicts. RCH was added because I encountered it in
tests. RSC normally should never occur in the sales book case because the
XML is validated in advance, but it's listed in the spec as a possible
result so I included it as well. The different lists of response codes
across the duplicates implementations should be refactored at one point
to avoid the current differences.

In the case of _analyze_sii_sales_book_result, because the token being
invalid was only checked after the response had been analyzed, and
unexpected responses caused the SII status to become `rejected`,
the invalid token would not get cleared and the sales book would be
stuck in the `rejected` state. The cron job will only retry books in the
`ask_for_status` state. Like for l10n_cl_edi, checking for an invalid
token response before analyzing the response and clearing the token if
it was invalid fixes this. No retry was implemented like in l10n_cl_edi,
so if this happens the user (or cron job) needs to verify twice and
after that the SII status will update.

In the case of _analyze_sii_result_rest, if the token is invalid it will
be detected in _get_send_status_rest, which will retry indefinitely
until it succeeds. The response format for the boletas REST service is
also different. So no token code checking modifications were needed for
that. Aligning that with the other implementations is better suited for
a future refactoring task.

There was also a typo in the _analyze_sii_result_rest implementation
where the folio in the response was compared with the document type
on the invoice. This has been replaced in
https://github.com/odoo/enterprise/pull/22207 with a comparison on both
the folio and the document type. The check is there because multiple
documents can be returned in the same response in theory, although
in practice there's currently only 1 response.

closes odoo/enterprise#22391

X-original-commit: dbd9eea45d6f692ac6829b03498d72779dc9af40
Signed-off-by: Josse Colpaert <jco@openerp.com>