[FIX] industry_fsm: prevent tour to stop
This issue was raised as it was failing in the `Single App Test` Build.
The problem comes from the fact that the selector used in order to go
back to the task form view was actually triggering a back to the tasks
kanban view of the project.

This commit also remove a useless id in the steps.

Prior to this commit:

    * The tour was stopping at step `Confirm the <b>time spent</b> on your task. <i>
      Tip: a rounding of 15min has automatically been applied.</i>`.

After this test:

    * The tour will run smoothly up to the end.

task-2692191

closes odoo/enterprise#22274

Signed-off-by: Xavier <xbo@odoo.com>