[FIX] sale_amazon: unescape HTML escape sequences in string data
What's the issue?
When we sync orders placed on Amazon, we use the 'Title' field returned
by Amazon's API to populate the product description field in the
generated sale order (see amazon_account.py).

The data in this 'Title' field sometimes contains HTML escape sequences
(e.g. &#x308; for 'ä'), which are then used as-is in the product
description without being converted back to Unicode.

For example, this is an issue in support ticket no. 2637289. As you can
see in the screenshots posted by the customer, the escape sequence
&#x308; appears in the product description.

I'm not sure why these escape sequences were present in the data sent
by Amazon. Perhaps the client had inadvertently used these escape
sequences when defining the product in Seller Central?
However, the product page on Amazon does not contain these escape
sequences in its HTML.

What's the proposed solution?
Call Python's built-in html.unescape() on the text returned by Amazon.

closes odoo/enterprise#22679

X-original-commit: 83384bbf566137702130f6e983ef858022272925
Signed-off-by: Antoine Dupuis (andu) <andu@odoo.com>
Signed-off-by: Antoine Vandevenne (anv) <anv@odoo.com>