[IMP] planning,project_forecast: improve the user/validation errors
The purpose of this commit is, to improve the statements of
user/validation errors as sometimes they are incorrect in English
or obscure and can be confusing for the user to understand the issue.
Also some errors could be avoided beforehand to provide better UX.

In this commit, improve the user/validation error messages and improve
the behaviour of the task deletion, when we try to delete a task
containing shifts, instead of rasing error the task on shifts should be
set as fasle.

task-2615466

closes odoo/enterprise#22393

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>