[FIX] stock_barcode: traceback when click on apply
How to reproduce:
  - Create a product tracked by SN and set a barcode for this product;
  - Update the quantity for this product: 1 qty with no serial number
    (don't forget to apply);
  - Go to the barcode's "Inventory Adjustments" and scan this product;
    --> You should have one line for this product with 1/1 qty and with
        no serial number;
  - Click on the set button to set the line as modified (qty should now
    be 0/1);
  - Click on apply, a wizard to confirm you want to modify this quant
    with no serial number should appears. Cancel the wizard;
  - Click again on the set button, qty should now back to 1/1;
  - Click again on apply --> Traceback.

The traceback was due to the fact after a save, we get back the updated
records and we refresh them, meaning the existing lines are updated or
new ones are created if needed.
But when checking if a quant record corresponding to an existing line,
we actually make a strict compare between a string and a number for the
id, and so, the line is recreated instead of be updated, which mean the
next time there will be a save, it will try to find a line who doesn't
exist anymore.

task-2682383

closes odoo/enterprise#22989

X-original-commit: f0aef7a028a9e121496d0ca2ea1735c19948c8fc
Signed-off-by: Arnold Moyaux <arm@odoo.com>