[IMP] sale_planning: open SOL form view only when user is salesman
Before this commit, the SOL in the form view of `planning.slot` model
can be opened to see its form view. The problem is if the user has not
access rights to the Sales App, there is no reason to let this user to
see the form view of the SOL.

This commit lets to open the SOL in the form view in `planning.slot`
model to see the form view of the SOL only if the user has at least the
`group_sale_salesman` group.

task-2671848

Community PR: odoo/odoo#81027

Closes #22773