[FIX] project_account: fix the analytic accounting setting
Before this fix, the "Track costs & revenues by project" option is missing
from the project module if both project and invoicing are installed.

Now, this setting is shown on the project module until accounting is installed,
as expected. It is then shown in the settings of the account module.

task-2695812

closes odoo/enterprise#22370

Related: odoo/odoo#80045
Related: odoo/upgrade#3051
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>