[FIX] helpdesk_sale_timesheet: fix visibility of project invoicing tab
This commit changes the visibility condition of the invoicing tab in
project form view when `helpdesk_sale_timesheet` module is just
installed in the db. Because, when a project is linked to a helpdesk
team, the invoicing tab should be visible when the project is billable
and partner is set (same condition given in `sale_timesheet` module).

Steps to reproduce:
==================
1) install `helpdesk_sale_timesheet` module in empty database.
2) Create a helpdesk team
3) Enable "Timesheets" option in "Track & Bill Time" section and save.
4) Go to the project linked to the helpdesk team
5) Select a partner

Expected Behavior:
=================
The invoicing tab should be visible.

Actual Behavior:
===============
The invoicing tab is not visible.

task-2647164

closes odoo/enterprise#22413

X-original-commit: 9e56e28806836d6955b388ba1c0382219e1888a8
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>