[IMP] add setting to prevent an user to encode old timesheets
Currently, nothing prevents users from encoding timesheets way in the past.
This could easily go unnoticed by managers, if they only focus on the last week or month
when validating timesheets. This could be an issue as past timesheets can impact the
invoicing of services, as well as the overall time record of the employees.

The goal of this task is to add an option to prevent users from encoding timesheets after a certain number of days.

task-2695707

closes odoo/enterprise#22407

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>