[FIX] hr_gantt: name on overflow if employee avatar
In 14.0, with the structure of the title row on gantt view, if there is
a row with:

 [employee avatar] [employee name]

that has any overflow, we don't see the avatar or employee name and just
see ellipsis symbol (...).

Depending on side of window, columns, length of employee name this is
easy to get the issue.

With this changeset we clip overflow text in this particular case
(showing employee avatar) so we see the text but cut instead of
ellipsed.

opw-2671160

closes odoo/enterprise#22427

X-original-commit: 616e224781cee2d9b9c3d676c3b3ce0f64b7c0f8
Signed-off-by: Nicolas Lempereur (nle) <nle@odoo.com>