[FIX] mrp_workorder: barcode scan with multiple lots sharing name
Usecase to reproduce:
- Create a Lot1 for product A and B
- Add product A in an operation with a workorder step
- In tablet view when asking to enter lot name for product A scan the
Lot1

Current result:
Traceback since it try to add multiple record to a many2one

Expected result:
Lot 1 is set on the tablet view

It happens because on_barcode_scan doesn't check the product while
searching the barcode. It would even be possible to set a lot not
related to the product. Fix it by adding the product_id in the search

closes odoo/enterprise#22454

X-original-commit: 557c47628a585b81aa0c988e6e422db353bf9520
Signed-off-by: Arnold Moyaux <arm@odoo.com>