[FIX] web_studio: re-render main navbar when leaving studio
In studio, edit the menus in order to have overflowing menu items that should go into
a "More" dropdown.

Leave studio after that.

Before this commit, the main navbar after leaving did not have the more dropdown.

After this commit it has, since when leaving studio, we force it to re-render and the adapt

closes odoo/enterprise#22498

X-original-commit: b9d192067d289681c55948687ab6d3bc3a267d75
Related: odoo/odoo#80353
Signed-off-by: Géry Debongnie (ged) <ged@openerp.com>
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>