[FIX] l10n_cl_reports: fix values of Resultado del Ejercicio in Balance Tributario
- Install Chile localization
- Create an invoice and post it
- Go to Accounting > Reporting > Balance Tributario (8 columnas)
"Resultado del Ejercicio" and "Total" lines are incorrect.

In "Resultado del Ejercicio" line, the positive difference between Perdida subtotal
and Ganancia subtotal is put in the column with the greater value to show the difference
and is not taken into account in the Total.

It should be the opposite.
The "Resultado del Ejercicio" line should contain the positive difference between
Perdida subtotal and Ganancia subtotal in the column with the lower value to balance
both Perdida and Ganancia in Total line.

opw-267926

closes odoo/enterprise#22578

X-original-commit: ab4d95d9249c4ec2f07dccd923e10f350815a435
Signed-off-by: Josse Colpaert <jco@openerp.com>
Signed-off-by: Anh Thao PHAM <pta@odoo.com>