[FIX] sale_purchase_inter_company_rules: no trigger auto purchase
Auto purchase is product-based and not company based which is
problematic in the case of intercompany auto purchase of product from
one company to another.  An error is raised when a product is
auto-purchased without vendor, but as we cannot set auto purchase option
differently in each company, it may cause an issue when handling
multiple companies.

Solve : In the case of auto-purchase, we only purchase when the
transaction user is not the intercompany one set on the selling company

opw-2569192

closes odoo/enterprise#22527

X-original-commit: e0ffb101b5920f27e9f608a83c8ea50baa73e45f
Related: odoo/odoo#80416
Signed-off-by: William Henrotin (whe) <whe@odoo.com>