[FIX] account_reports: fix for cash flow statement report
Before this Commit:
===================
If account_journal_payment_debit_account_id and account_journal_payment_credit_account_id
is not set on any company then the cash flow statement is printed with zero values for all
the companies.

After this Commit:
==================
The cash flow statement is printed with correct values for all the companies.

opw - 2671338
opw - 2671177

closes odoo/enterprise#22495

X-original-commit: 3c8751de21fc06ec88a839bb54872a37ea917ffc
Signed-off-by: Quentin De Paoli (qdp) <qdp@openerp.com>
Signed-off-by: Florian Gilbert <flg@odoo.com>