[FIX] industry_fsm_sale: remove useless company context key
This commit removes the use of the default_company_id context key as
the related PR odoo/odoo#79926 sets the company_id of the invoices
generated through the sale.advance.payment.inv wizard to the company_id
of the SO it has been generated from.

This commit also re-introduce multicompany tests that had been disabled
in order to ensure that the multicompany use case is respected.

task-2608812
Closes: #20514
Related PR: odoo/odoo#79926