[IMP] services: convert tickets to tasks and tasks to tickets
In some cases, a ticket can lead to the creation of tasks. For instance:
- the customer submitted a suggestion/feedback that is forwarded to the R&D
- the ticket needs further investigation from an expert team, outside of the support

It can also be the case the other way around. For example, a customer might report a weird behavior to the R&D, that turns out to be a bug that needs to be forwarded to the helpdesk.

Currently, the user has to manually create the record and copy all the information by hand. This is annoying and time consuming.

This commit adds an action to directly convert a ticket into a task, and
vice versa.

Task 2687048

closes odoo/enterprise#22531

Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>