[REF] documents_spreadsheet_bundle: simplify tests
The goal of this commit is to remove the complexity of the helpers of the
testst (createSpreadsheetFromList, createSpreadsheetFromPivot, ...)

The new helpers use a webclient configured with everything needed for
spreadsheet (registries, collaborative, ...), and use the UI to create a
list or a pivot. There are two main advantages of this: it's simplier and
more closer from reality.

This commit also move tests files the correct folder

Part of task-id 2658217

Part-of: odoo/enterprise#22487