[REM] l10n_cl_edi: Remove bugged _get_partner stable code path
In https://github.com/odoo/enterprise/pull/21956 a bugfix was
implemented for _get_partner. It was matching on partners that didn't
belong the company that would be set on the invoice, resulting in an
exception while saving the record and read access errors when accessing
the invoice in the UI.

The fix changed the function signature, and the old code path with the
bug in it was kept to keep the signature stable. This commit for master
removes that old code path.

closes odoo/enterprise#22503

Signed-off-by: Josse Colpaert <jco@odoo.com>