[FW][FIX] appointment: use the event_tz in mail templates
When sending a mail for an appointment the time displayed depends
on the user timezone. Indeed no timezone was specifically used for
the datetime. This was leading to different time in the mail sent
based on the timezone of the browser of the user that was sending
it.

To avoid confusion, we will now displayed the timezone of the
appointment type in the mail and adapt the datetime with this
timezone. We set the event_tz of the event created with the
appointment type timezone.

We also display the timezone used by the user when he selects his slot
before taking an appointment on the frontend

task-2699765

closes odoo/enterprise#25322

Part-of: odoo/enterprise#22599
X-original-commit: efa7f79e737486ce84b9b55b74bc0c47928d925a
Related: odoo/odoo#86513
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>