[FIX] l10n_sa_hr_payroll: remove typo in payroll rule and change category of social contribution rule
A typo causing an error in the rule 'Expat End of Service Benefit'. The rule 'Social Insurance Contribution' was in the category deduction but that causes it to be removed from net salary. It is now in the category 'company contribution'. Remove the no update to allow easy update in the future in case of changes (like a law change)

closes odoo/enterprise#22529

X-original-commit: 6a00bb38a108d4b8a30b978d201f155c67a328d9
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>