[FIX] hr_payroll: Display 'Has negative net to report' checkbox
Let's take the following example:

MARCH: Jules receives 2000 €
APRIL:
- We make a refund of -2000 € (we cancel the fdp what)
- We make a new sheet of 2300 € (It's the right yip)

For the moment, the SEPA file for April will contain an amount
of € 2,300 for Jules (as the negative amounts are excluded),
It was proposed to make an adaptation so that the file contains
only € 300 (the difference).

This is not really correct, because the -2000€ record is
going to be marked as "Negative net to report"

And in fact, I think we already have everything in hand. We should :
- Transfer the -2000€ to the new 2300 sheet (the net will then be 300).
- Uncheck the "negative net to report" box on the -2000€ form after
  validation of the batch (since it is already reported).

As the negative net to report box is hidden, this commit displays it
to allow managing this use case.

Otherwise we will do some magic stuff and reduce the UX aspect (Why such
and such thing is marked as net to report or not, and also if
we validate several batches at the same time, it will be a mess).

closes odoo/enterprise#22562

X-original-commit: 5c88b9b94ae3e9c7f38226ac1538e19962537b91
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>