[IMP] helpdesk: add group to hide ratings feature
Before this commit, since we have no group, we cannot hide the filters
about customer ratings feature in helpdesk.ticket when no team has
`use_rating` enabled. In the Project app, we can do it because it is a
global setting that add a group to the internal user.

This commit adds a group to the helpdesk user when at least a helpdesk
team has a "Customer rating" feature enabled and also remove this group
when no team has the feature enabled. With this group, we are able to
hide the filters for customer ratings when the feature is not enabled in
all helpdesk teams.

task-2671848

Community PR: odoo/odoo#81027

Closes #22773