[IMP] helpdesk: avoid remove use_sla_group if it is done
Before this commit, when no team has SLA feature enabled and the current
user creates a new team, we will do a write in the helpdesk user group
to remove the `use_sla_group` but this group should not be in the
helpdesk user group since no team has the feature enabled.

This commit avoids doing a write each time, we check if the user has the
group and if no helpdesk team has the feature to remove the group in
helpdesk user.

task-2671848

Community PR: odoo/odoo#81027

Closes #22773