[FIX] hr_payroll: Prevent re-computing a refund
The goal of a refund is to "cancel" a payslip after it has been
posted in the accounting entries.

It makes no sense to modify it afterward, and re-computing it
is gonna reset some rules (Basic for example), et could introduce
unwanted side effects in the rule computations.

So create the refund in Waiting state, and hide the button.

X-original-commit: 5267b4c92c8292f532bf88f64c326cb6919635fd
Part-of: odoo/enterprise#22566