[FIX] account_invoice_extract: fix auto-send of customer invoices
The automatic sending of customer invoices to the extraction server
will only happen if it has been created via the email alias or via
the 'Upload' button.

Most customer invoices will not have its corresponding PDF/image
document attached as it is being created with Odoo. It won't
need the use of the extraction server. This will prevent unrelated
attachments from being sent and analysed.

closes odoo/enterprise#22684

X-original-commit: ed34f58c96f391b99ede9b8cf3960769225f2b1b
Signed-off-by: William André (wan) <wan@odoo.com>
Signed-off-by: Louis Baudoux (lba) <lba@odoo.com>