[FIX] web_enterprise: stop expiration panel unexpectedly unblocking UI
Previously, the expiration panel would unblock the UI in a bunch of
flows even when it did not block it previously, this means that it can
unblock the UI that was blocked by unrelated code. This commit makes it
so that the expiration panel only unblocks the UI if it blocked it
previously.

closes odoo/enterprise#22590

X-original-commit: b940857b55f6f7e1cc91ff9244eb3d01b4ab98a2
Related: odoo/odoo#80543
Signed-off-by: Lucas Perais (lpe) <lpe@odoo.com>
Signed-off-by: Samuel Degueldre <sad@odoo.com>