[FIX] planning: correct compute of start/end dates of planning shifts
When computing start/end dates of shifts, the static method _calculate_start_end_dates
is used and is calling _get_tz method. In case of a recomputation on several records,
this causes an issue as this latter method is using self values, so it could generate
incorrect results or a traceback if several employees are defined in the recordset.

This commit also fixes an issue in the planning tests. From this commit dc67b3278f5cfcd0c88ed35a74115a776c31189d,
the employee_id has become a compute field from the ressource_id. Some tests were not
addapted in this PR #19695, which this commit fixes.

opw-2700259

closes odoo/enterprise#22645

X-original-commit: dff238bc62018ffd4f65a007b45fa2e3c93099c3
Signed-off-by: Laurent Stukkens (ltu) <ltu@odoo.com>