[IMP] sign: send mail automatically for new cc partners after signed
when a cc partner is added to a 'signed' sign request
1. the completed documents and message_cc will be sent to the ccer
2. a Log Note will be created in Chatter
2. the cc partner will be added to the followers
Note: removing a ccer for a sign request will not remove him from the follower list automatically.

before this commit:
adding cc partner for a 'signed' sign request cannot send completed documents to the new cc partner automatically.

Taskid: 2659342
Part-of: odoo/enterprise#22813