[FIX] l10n_sa_hr_payroll_account: check coa in company
During auto-installation of this module, the check below raises an error.
https://github.com/odoo/enterprise/blob/3c962644c78f67b9b540c5f80a1c4667d6955d06/l10n_sa_hr_payroll_account/models/account_chart_template.py#L40-L41

The reason is that these accounts are not present even if the company
country is SA.
https://github.com/odoo/enterprise/blob/3c962644c78f67b9b540c5f80a1c4667d6955d06/l10n_sa_hr_payroll_account/models/account_chart_template.py#L21-L31

Since the xmlid of the SA CoA was renamed from `account_arabic_coa_general`
https://github.com/odoo/odoo/blob/3ae0b21f4b269859dc41ab6851e2e6e52b6b2e11/addons/l10n_sa/data/account_chart_template_data.xml#L3
to `sa_chart_template_standard`
https://github.com/odoo/odoo/blob/6e68df1a69079bc16888ad588c97f234c9b9fa07/addons/l10n_sa/data/account_chart_template_data.xml#L3
and the account templates also changed completely
https://github.com/odoo/odoo/pull/77673/files#diff-57b148b690d198d5579f14ac190846b9f26f551d335f5cb57e9395cc155afdaeL1

This is preventing the upgrade to 15.0 of almost all SA localized DBs.

The fix we propose is to check if the CoA of the company is the expected
one, `sa_chart_template_standard`, before attempting to setup the
accounts.

closes odoo/enterprise#22803

X-original-commit: f775a1ed137f8a0c0eef3ca28b3571f5b35098d7
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>
Co-authored-by: Josse Colpaert <jco@openerp.com>