[REF] web_cohot, *: remove dropdown-item-selected event
*: web_enterprise, web_studio, documents_spreadsheet

In owl 2, the t-on directive is no longer supported on components. The
way to handle events on components is to give them a callback as a prop.
This already works in owl 1, so this commit prepares the code for an
easier migration in the future by removing the "dropdown-item-selected"
event and converting the code that was using it to use the new paradigm.

closes odoo/enterprise#22722

Related: odoo/odoo#80684
Signed-off-by: Géry Debongnie <ged@odoo.com>