[IMP] social_[twitter|demo]: adds support for retweets and quotes
To retweet or quote a given tweet, the user has to go on Twitter, log in
with the account of the company and retweet or quote the given tweet from
Twitter. This current flow is not really convenient as the user has to
leave the app and may have to provide some credentials to log in.

With this new commit, the user will now be able to retweet and quote a
tweet directly from Odoo. The interface will be similar to Twitter: The
quoted tweet will be represented in an embeded view and the tweet cards
will have a small dropdown menu allowing the user to retweet or quote a
tweet.

The user will have the same limitations as on Twitter: The user will
not be allowed to retweet a tweet twice with the same account but the
user will be allowed to quote a tweet multiple times with the same account.

task-2518289

closes odoo/enterprise#20660

Related: odoo/upgrade#3082
Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>